+++ packages/platform-sdk-ark/__tests__/services/peer.test.ts
@@ -23,7 +23,7 @@ describe("PeerService", () => {
 				});
 
 				const peerService: PeerService = await PeerService.construct({
-					peer: "http://127.0.0.1",
+					peer: "http://127.0.0.1/api",
 				});
 
 				expect(peerService.getSeeds()).toEqual(
@@ -40,7 +40,7 @@ describe("PeerService", () => {
 				});
 
 				const peerService: PeerService = await PeerService.construct({
-					peer: "http://127.0.0.1",
+					peer: "http://127.0.0.1/api",
 				});
 
 				expect(peerService.getSeeds()).toEqual(
@@ -58,7 +58,7 @@ describe("PeerService", () => {
 
 				await expect(
 					PeerService.construct({
-						peer: "http://127.0.0.1",
+						peer: "http://127.0.0.1/api",
 					}),
 				).rejects.toThrowError(new Error("No seeds found"));
 			});
@@ -70,7 +70,7 @@ describe("PeerService", () => {
 					.get("/mainnet.json")
 					.reply(200, dummySeeds);
 
-				const peerService: PeerService = await PeerService.construct({ network: "mainnet" });
+				const peerService: PeerService = await PeerService.construct({ network: "live" });
 
 				expect(peerService.getSeeds()).toEqual(dummySeeds.map((peer) => ({ ip: peer.ip, port: 4003 })));
 			});
@@ -90,7 +90,7 @@ describe("PeerService", () => {
 
 				await expect(
 					PeerService.construct({
-						network: "mainnet",
+						network: "live",
 					}),
 				).rejects.toThrowError(new Error("No seeds found"));
 			});
@@ -104,7 +104,7 @@ describe("PeerService", () => {
 				data: dummyPeersWalletApi,
 			});
 
-			peerService = await PeerService.construct({ peer: "http://127.0.0.1" });
+			peerService = await PeerService.construct({ peer: "http://127.0.0.1/api" });
 		});
 
 		it("should find peers", async () => {
@@ -191,7 +191,7 @@ describe("PeerService", () => {
 				data: dummyPeersWalletApi,
 			});
 
-			peerService = await PeerService.construct({ peer: "http://127.0.0.1" });
+			peerService = await PeerService.construct({ peer: "http://127.0.0.1/api" });
 		});
 
 		it("should find peers without the wallet api plugin", async () => {
@@ -245,7 +245,7 @@ describe("PeerService", () => {
 				data: dummyPeersWalletApi,
 			});
 
-			peerService = await PeerService.construct({ peer: "http://127.0.0.1" });
+			peerService = await PeerService.construct({ peer: "http://127.0.0.1/api" });
 		});
 
 		it("should find peers without estimates", async () => {